// src/generated/models/Craft_clientsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Craft_clientscraft_clientstatus = {
  0: 'Active',
  1: 'Inactive',
  2: 'Not in Contact',
  3: 'Blacklisted'
} as const;
export type Craft_clientscraft_clientstatus = keyof typeof Craft_clientscraft_clientstatus;
export const Craft_clientsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Craft_clientsstatecode = keyof typeof Craft_clientsstatecode;
export const Craft_clientsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Craft_clientsstatuscode = keyof typeof Craft_clientsstatuscode;

export interface Craft_clientsBase {
  craft_activeprojectscount?: number;
  craft_clientaddress?: string;
  craft_clientid: string;
  craft_clientid1?: string;
  craft_clientname?: string;
  craft_clientstatus?: Craft_clientscraft_clientstatus;
  craft_clienttags?: string;
  craft_clienttype?: string;
  craft_contactdesignation?: string;
  craft_contactemail?: string;
  craft_contactperson?: string;
  craft_contactphone?: string;
  craft_creditlimitaed?: number;
  craft_internalnotes?: string;
  craft_isactive?: boolean;
  craft_lastactivitydate?: string;
  craft_lifetimerevenue?: number;
  craft_paymentterms?: string;
  craft_recordcreatedat?: string;
  craft_recordcreatedby?: string;
  craft_recordlastupdatedat?: string;
  craft_recordversion?: number;
  craft_riskprojectscount?: number;
  craft_taxregistrationnumber?: string;
  craft_tradelicensenumber?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Craft_clientsstatecode;
  statuscode?: Craft_clientsstatuscode;
  timezoneruleversionnumber?: number;
  "transactioncurrencyid@odata.bind"?: string;
  utcconversiontimezonecode?: number;
}

export interface Craft_clients extends Craft_clientsBase {
  craft_clientstatusname?: string;
  craft_creditlimitaed_base?: number;
  craft_isactivename?: string;
  craft_lifetimerevenue_base?: number;
  createdon?: string;
  exchangerate?: number;
  modifiedon?: string;
  ownerid: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}