// src/generated/models/Craft_projectsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Craft_projectscraft_healthstatus = {
  0: 'Amber',
  1: 'Green'
} as const;
export type Craft_projectscraft_healthstatus = keyof typeof Craft_projectscraft_healthstatus;
export const Craft_projectsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Craft_projectsstatecode = keyof typeof Craft_projectsstatecode;
export const Craft_projectsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Craft_projectsstatuscode = keyof typeof Craft_projectsstatuscode;

export interface Craft_projectsBase {
  craft_budgetvariancepercentage?: number;
  "craft_client@odata.bind"?: string;
  craft_clientid?: string;
  craft_completionpercentage?: number;
  craft_contractvalueaed?: number;
  craft_createdat?: string;
  craft_createdby?: string;
  craft_dmpermitnumber?: string;
  craft_durationdays?: number;
  craft_elapseddays?: number;
  "craft_employeerecord@odata.bind"?: string;
  craft_enddate?: string;
  craft_grossfloorareasqm?: number;
  craft_healthstatus?: Craft_projectscraft_healthstatus;
  craft_internalnotes?: string;
  craft_isactive?: boolean;
  craft_lastupdatedat?: string;
  craft_location?: string;
  craft_numberoffloors?: number;
  craft_plotnumber?: string;
  craft_profit?: number;
  craft_projectdescription?: string;
  craft_projectid: string;
  craft_projectid1?: string;
  craft_projectmanagerid?: string;
  craft_projectname?: string;
  craft_projecttype?: string;
  craft_recordversion?: number;
  craft_remainingdays?: number;
  craft_riskscore?: number;
  craft_startdate?: string;
  craft_status?: string;
  craft_tags?: string;
  craft_totalcost?: number;
  craft_totalrevenue?: number;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Craft_projectsstatecode;
  statuscode?: Craft_projectsstatuscode;
  timezoneruleversionnumber?: number;
  "transactioncurrencyid@odata.bind"?: string;
  utcconversiontimezonecode?: number;
}

export interface Craft_projects extends Craft_projectsBase {
  craft_contractvalueaed_base?: number;
  craft_healthstatusname?: string;
  craft_isactivename?: string;
  craft_profit_base?: number;
  craft_totalcost_base?: number;
  craft_totalrevenue_base?: number;
  createdon?: string;
  exchangerate?: number;
  modifiedon?: string;
  ownerid: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  craft_client?: object;
  _craft_client_value?: string;
  craft_employeerecord?: object;
  _craft_employeerecord_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}